added adaptive stroke flag; modified demo to show new features; fixed eslint complaint
updated readme

Update README.md
