+++ demo/index.html
@@ -39,16 +39,22 @@
 				font-family: sans-serif;
 				padding: 1em;
 			}
+
+			#clear {
+				display: none;
+			}
 		</style>
 	</head>
 	<body>
 		<h1>atrament.js</h1>
 		<form>
-			<button onclick="event.preventDefault(); atrament.clear();">clear</button><br>
+			<button id="clear" onclick="event.preventDefault(); atrament.clear();">clear</button><br>
 			<label>Thickness</label><br>
 			<input type="range" min="1" max="40" oninput="atrament.weight = parseFloat(event.target.value);" value="2" step="0.1" autocomplete="off"/><br>
 			<label>Smoothing</label><br>
 			<input type="checkbox" onchange="atrament.smoothing = event.target.checked;" checked autocomplete="off"><br>
+			<label>Adaptive stroke</label><br>
+			<input type="checkbox" onchange="atrament.adaptiveStroke = event.target.checked;" checked autocomplete="off"><br>
 			<label>Mode</label>
 			<select onchange="atrament.mode = event.target.value;">
 				<option value="draw" default>Draw</option>
@@ -60,10 +66,16 @@
 			<label>Opacity</label><br>
 			<input type="range" min="0" max="1" onchange="atrament.opacity = parseFloat(event.target.value);" value="1" step="0.05" autocomplete="off">
 		</form>
-		<canvas></canvas>
+		<canvas id="sketcher"></canvas>
 		<a href="https://github.com/jakubfiala/atrament.js"><img style="position: fixed; top: 0; right: 0; border: 0; z-index: 3;" src="https://camo.githubusercontent.com/38ef81f8aca64bb9a64448d0d70f1308ef5341ab/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f6769746875622f726962626f6e732f666f726b6d655f72696768745f6461726b626c75655f3132313632312e706e67" alt="Fork me on GitHub" data-canonical-src="https://s3.amazonaws.com/github/ribbons/forkme_right_darkblue_121621.png"></a>
 	</body>
 	<script>
-		var atrament = atrament('canvas', window.innerWidth, window.innerHeight);
+		var canvas = document.getElementById('sketcher');
+		var atrament = atrament(canvas, window.innerWidth, window.innerHeight);
+
+		var clearButton = document.getElementById('clear');
+		canvas.addEventListener('dirty', function(e) {
+			clearButton.style.display = atrament.dirty ? 'inline-block' : 'none';
+		});
 	</script>
 </html>
